new hook useStaticPreview
